refactor: extract APIKeyPersistenceState.swift from SettingsStore.swift
Extracts the `APIKeyPersistenceState` enum (persistence status for API
keys) from the bottom of SettingsStore.swift into its own file.

Byte-identical move of lines 1951-1957. New file needs only Foundation
(SettingsStore.swift keeps Combine + Foundation).

SettingsStore.swift: 1963 → 1954 lines. Tests: 667.
