+++ Sources/BugNarrator/Services/APIKeyPersistenceState.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+enum APIKeyPersistenceState: Equatable {
+    case empty
+    case keychain
+    case keychainLocked
+    case sessionOnly
+    case pendingSave
+}
